+++ src/components/layout/Footer.js
@@ -6,7 +6,7 @@ import vsethLogo from '../../assets/vseth_Logo_bylines_organisation-white.svg'
 import {makeStyles} from '@material-ui/core/styles';
 import Container from '@material-ui/core/Container';
 import Link from '@material-ui/core/Link';
-import Aux from "../../hoc/Aux";
+import Aux from "../../hoc/Au_x";
 import constants from '../../common/constants';
 
 function Copyright() {
