+++ src/components/pages/content/ContentRender.js
@@ -2,7 +2,7 @@ import React from 'react';
 import useMediaQuery from "react-responsive/src/useMediaQuery";
 import ReactMarkdown from "react-markdown";
 import Grid from '@material-ui/core/Grid';
-import Aux from "../../../hoc/Aux";
+import Aux from "../../../hoc/Au_x";
 import FbPagePlugin from "./FbPagePlugin";
 // import QuickLinks from "./QuickLinks";
