+++ src/hoc/Au_x.js
@@ -1,3 +1,3 @@
 const aux = props => props.children;
 
-export default aux;
+export default aux;
